--- src/components/MaanedsbloepAvansertBeregning/MaanedsbeleopBoksMobil/index.ts
@@ -1,1 +1,0 @@
-export { MaanedsbeleopBoksMobil } from './MaanedsbeleopBoksMobil'
