+++ src/components/MaanedsbloepAvansertBeregning/MaanedsbeloepMobil/index.ts
@@ -1,0 +1,1 @@
+export { MaanedsbeloepMobil } from './MaanedsbeloepMobil'
