+++ src/components/MaanedsbloepAvansertBeregning/MaanedsbloepAvansertBeregning.module.scss.d.ts
@@ -1,0 +1,5 @@
+declare const classNames: {
+  readonly maanedsbeloepMobileOnly: "maanedsbeloepMobileOnly";
+  readonly maanedsbeloepDesktopOnly: "maanedsbeloepDesktopOnly";
+};
+export = classNames;
